+++ src/offline/TFIDF_Tuple.java
@@ -8,9 +8,9 @@ package offline;
 // Data structure for storing multiple pieces of term-related information
 public class TFIDF_Tuple {
 	
-	public float tf_value;
-	public float idf;
-	public float tfidf_value;
+	public double tf_value;
+	public double idf;
+	public double tfidf_value;
 	public String word;
 	public String author;
