+++ src/online/CosSimilarityDenomCombiner.java
@@ -9,14 +9,14 @@ public class CosSimilarityDenomCombiner  extends Reducer<Text,Text,Text,Text>{
 	
 	public void reduce(Text  key,  Iterable<Text>  values,  Context  context) throws IOException, InterruptedException {
 
-		float sq_sum = (float) 0.0;
+		double sq_sum = 0.0;
 		for (Text val: values){
 			String[] split = val.toString().split("\t");
 			String tfidf = split[0];
-			float tfidf_value = Float.parseFloat(tfidf);
+			double tfidf_value = Double.parseDouble(tfidf);
 			sq_sum += (tfidf_value * tfidf_value);
 		}
-		float root_sq_sum = (float) Math.sqrt(sq_sum);
+		double root_sq_sum = Math.sqrt(sq_sum);
 		context.write(key, new Text(root_sq_sum + ""));
 	}
